// packages/study-echarts/src/echarts/util/types.ts
interface PayloadItem {
  excludeSeriesId?: OptionId | OptionId[];
  animation?: PayloadAnimationPart
  // TODO use unknown
  [other: string]: any;
}

export interface Payload extends PayloadItem {
  type: string;
  escapeConnect?: boolean;
  batch?: PayloadItem[];
}

export interface HighlightPayload extends Payload {
  type: 'highlight';
  notBlur?: boolean
}

export interface DownplayPayload extends Payload {
  type: 'downplay';
  notBlur?: boolean
}

// Payload includes override anmation info
export interface PayloadAnimationPart {
  duration?: number
  easing?: AnimationEasing
  delay?: number
}

export interface SelectChangedPayload extends Payload {
  type: 'selectchanged'
  escapeConnect: boolean
  isFromClick: boolean
  fromAction: 'select' | 'unselect' | 'toggleSelected'
  fromActionPayload: Payload
  selected: {
    seriesIndex: number
    dataType?: SeriesDataType
    dataIndex: number[]
  }[]
}
// ...
export type SeriesDataType = 'main' | 'node' | 'edge';
// ...
export type OptionId = string | number;